+++ scripts/release.py
@@ -104,6 +104,11 @@ def lvgl_update_library_json(v):
 	f.write(outbuf)
 	f.close()
 
+def lvgl_update_lv_conf_templ(ver_str):
+	title("lvgl: Update version number in lv_conf_template.h")
+	cmd("sed -i -r 's/v[0-9]+\.[0-9]+\.[0-9]+/"+ ver_str +"/' lv_conf_template.h ")
+
+
 def lvgl_commit_push(v):
 	title("lvgl: commit and push release")
 
@@ -235,7 +240,8 @@ lvgl_clone()
 lvgl_format()
 lvgl_update_api_docs()
 ver_str = lvgl_update_version()    
-lvgl_update_library_json(ver_str)
+lvgl_update_library_json(ver_str)  
+lvgl_update_lv_conf_templ(ver_str)
 lvgl_commit_push(ver_str)
 lvgl_merge_to_release_branch(ver_str)